requirements.txt
numpy
tqdm
fastapi
jinja2
uvicorn
pillow
pydantic
